backup tor RSA identity keys
in addition to the ed25519 key, we also need to backup the RSA identity
key to preserve node identities across reboots.
As TPM's NV indices are size-constrained we need to re-work the backup
format to fit as much data as possible:
- use bincode instead of JSON
- store relay info by id and not by name
- strip ed25519 key headers
- store RSA p,q, e components and regenerate PEM on restore

Follow up:
- use multiple NV indices if the backup exceeds the index max size
- opt-in network backup if the backup exceeds the sum of the NV indices
  size

However, from testing we should be able to store backups for up to 10
relay in a 2048 NV index, which seems reasonable.

to upgrade to the new backup format run `patela-client` on a live system
with `--skip-restore`.